App/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dnspython==2.7.0
futures==3.0.5
idna==3.10
ipaddress==1.0.23
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
platformdirs==4.3.6
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-whois==0.9.5
requests==2.32.3
rich==13.9.4
six==1.17.0
textual==3.1.1
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.4.0
